Refactor: Separate API route handlers into individual controllers
I've moved all API route handlers from `index.ts` into their own dedicated controller files within the `src/controllers/` directory. Each controller now manages a single API endpoint.

This change addresses the issue of `index.ts` being too long and improves code organization by separating concerns. `index.ts` now primarily handles middleware setup, server initialization, and mounting of the individual controller routers.

The following controllers were created:
- joinPostController.ts
- channelJoinPostController.ts
- getstreamWebhooksPostController.ts
- webhookPostController.ts
- attendancePostController.ts
- attendanceGetController.ts
- sendAttendanceMessagePostController.ts
- checkMessageStatusGetController.ts
- projectsGetController.ts

During the process, I identified and resolved TypeScript compilation errors and Jest test setup issues. All existing unit tests pass with these changes.
