+++ nodejs-ai-assistant/tsconfig.json
@@ -5,9 +5,11 @@
     "moduleResolution": "Node",
     "sourceMap": true,
     "outDir": "./dist",
+    "rootDir": "./src", // Added rootDir
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true
-  }
+  },
+  "include": ["./src/**/*"] // Specify what to compile
 }
